git: move get_refs_containing to backend
