+++ dvc/scm/git/backend/gitpython.py
@@ -32,6 +32,23 @@ class GitPythonBackend(BaseGitBackend):  # pylint:disable=abstract-method
     def remove_ref(self, name: str, old_ref: Optional[str] = None):
         raise NotImplementedError
 
+    def get_refs_containing(self, rev: str, pattern: Optional[str] = None):
+        from git.exc import GitCommandError
+
+        try:
+            if pattern:
+                args = [pattern]
+            else:
+                args = []
+            for line in self.git.for_each_ref(
+                *args, contains=rev, format=r"%(refname)"
+            ).splitlines():
+                line = line.strip()
+                if line:
+                    yield line
+        except GitCommandError:
+            pass
+
     def push_refspec(self, url: str, src: Optional[str], dest: str):
         raise NotImplementedError
